iot_integration/requirements.txt
certifi==2021.10.8
elasticsearch==7.15.2
paho-mqtt==1.6.1
urllib3==1.26.7
